Windows Form Example
'Your client requires an application to determine the age of the oldest
person in any family.
Your program will use a textbox to accept ages and add the numbers to a
list box.
'o find the oldest member of a family the user will click a button which
will use a loop to find the largest number in the list box.